+++ src/index.ts
@@ -1,3 +1,5 @@
+export { DefaultPubSubTopic } from "./lib/constants";
+
 export {
   generatePrivateKey,
   generateSymmetricKey,
@@ -12,7 +14,7 @@ export * as enr from "./lib/enr";
 export * as utils from "./lib/utils";
 
 export * as waku from "./lib/waku";
-export { Waku, DefaultPubSubTopic, Protocols } from "./lib/waku";
+export { Waku, Protocols } from "./lib/waku";
 
 export * as waku_message from "./lib/waku_message";
 export { WakuMessage } from "./lib/waku_message";
@@ -25,7 +27,7 @@ export {
 } from "./lib/waku_light_push";
 
 export * as waku_relay from "./lib/waku_relay";
-export { WakuRelay, RelayCodecs } from "./lib/waku_relay";
+export { WakuRelay } from "./lib/waku_relay";
 
 export * as waku_store from "./lib/waku_store";
 export { PageDirection, WakuStore, StoreCodecs } from "./lib/waku_store";
