+++ src/lib/constants.ts
@@ -1,0 +1,9 @@
+/**
+ * DefaultPubSubTopic is the default gossipsub topic to use for Waku.
+ */
+export const DefaultPubSubTopic = "/waku/2/default-waku/proto";
+
+export enum StoreCodecs {
+  V2Beta3 = "/vac/waku/store/2.0.0-beta3",
+  V2Beta4 = "/vac/waku/store/2.0.0-beta4",
+}
