+++ src/lib/waku_relay/get_relay_peers.ts
@@ -1,7 +1,7 @@
 import Gossipsub from "libp2p-gossipsub";
 import { shuffle } from "libp2p-gossipsub/src/utils";
 
-import { RelayCodecs } from "./index";
+import { RelayCodecs } from "./constants";
 
 /**
  * Given a topic, returns up to count peers subscribed to that topic
