+++ src/lib/waku_store/constants.ts
@@ -1,0 +1,4 @@
+export enum StoreCodecs {
+  V2Beta3 = "/vac/waku/store/2.0.0-beta3",
+  V2Beta4 = "/vac/waku/store/2.0.0-beta4",
+}
