+++ src/lib/waku_store/history_rpc.ts
@@ -5,7 +5,7 @@ import { v4 as uuid } from "uuid";
 import * as protoV2Beta3 from "../../proto/waku/v2/store/v2beta3/store";
 import * as protoV2Beta4 from "../../proto/waku/v2/store/v2beta4/store";
 
-import { StoreCodecs } from "./index";
+import { StoreCodecs } from "./constants";
 
 export enum PageDirection {
   BACKWARD = "backward",
